moves modifier base to its own file
